refactor(auth): 로그인한 유저의 정보를 담는 LoginMember를 member에서 auth로 이동
