+++ src/main/java/nextstep/subway/auth/domain/LoginMember.java
@@ -1,5 +1,7 @@
-package nextstep.subway.member.domain;
+package nextstep.subway.auth.domain;
 
+
+import nextstep.subway.member.domain.Member;
 
 public class LoginMember {
     private Long id;
