+++ src/test/java/nextstep/subway/auth/ui/interceptor/authorization/TokenSecurityContextPersistenceInterceptorTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.auth.ui.interceptor.authorization;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
 import nextstep.subway.auth.infrastructure.SecurityContextHolder;
-import nextstep.subway.member.domain.LoginMember;
+import nextstep.subway.auth.domain.LoginMember;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
